Hide buttons for inactive workspaces/projects
If a workspace is archived or its project is inactive, just hide
all the buttons on the workspace and request page (for the request
author). Nothing can be actioned by an author when a workspace is
in an inactive state.

We don't hide request buttons for output checkers (unless they are
the author), because we don't get project information for
output checkers.