Apply visibility rules to file votes
The visibility of these is tied to the state of the request, which
get_current_request_visibility() handles.

I refactored the way the file browser api gets the request status, from
needing two methods to just one, as they both need the user now.
Renamed some of the functions for clarity too.

The new combinded get_request_file_status() function takes a relpath and
a user, and returns the status for that path that that user is allowed
to see.

For authors, that is always INCOMPLETE until the request is RETURNED or
APPROVED.

For output checkers, they can always see their own vote, and they can
see the consolidated vote if the request is no longer BLINDED.