Show file votes based on visibility in the file view.
The replaces the logic that was based on urls, and instead bases it on
the get_request_file_status() logic, so we know the file status we have
is one the user is supposed to see.

It now displays the consolidated request status that you are currently
allowed to see, and if you are an output checker, it shows you what you
voted. Because of the incorrect logic previously, it had merged these
ideas.

Also refactored the way statuses are shown, using semantic classes and
a description() helper, although there is a plan afoot with Thomas to
switch this to data- attributes instead.

This UI stil has issues - you can't easily see who voted on the issue,
but it's better than what we have currently. I think there will be
a polish PR following shortly.