UI for comment visibility.
For output-checkers, who can see PRIVATE comments and their own BLINDED
comments, we highlight what the current visibility of the comment is. We
add a pill, and also provide a clear backgroun colour to help
understand. The actualy styles can be latter finessed.

To get coverage of the template branch, I ended up adding a some more
tests which I realised were missing, and extending the current test to
cover the comment visibility.