Rename file status enums
There were lots of `..Status` types, and it was getting noisy and
difficult to read.  This defines and uses some specific terms to try
make this easier to understand.

We use the term "vote" to refer a single output-checkers review of file.
We've been using this term informally in discussions - this change
formalises that. So UserFileReviewStatus is renamed to RequestFileVote.

We use the term "decision" to refer to the corporate status of a file,
based on all output-checker's individual votes. So
RequestFileReviewStatus becomes RequestFileDecision.

In theory, this leaves the term "review" free mean the entire review an
output checker peforms. We don't have an object or status for this, but
we do already use it in functions names, e.g. `complete_review`.