Update css classes for combined file decision status
In order to determine whether to show the user-decision or
combined decision, we need to know if the request is in an
independent stage of review (i.e. submitted or partially
reviewed). The path item doesn't know about the request status (it
confusingly has a request status but it's a FileReviewRequestStatus
not the status of the request itself). I've done this as a
data attribute (but have also removed some other data attributes that
I dont think were required from the CSS)