Refactor independent visibility implementation, using turns
Rather than tracking and mutating in the individual visibility of
comments, we instead introduce the concept of a centralised review
"turn", which is an integer that increments every time the request gets
transferred back and forth between author and output checkers.

Comments (and in future audit logs) store which turn they were created
in. We then use the this in the logic to show or hide comments based on
the current request phase and turn. Users can always see comments from
previous turns (only output-checkers can see private comments). Comments
from the current turn are subject to blinding as per independent review
process.

This is better that the previous solution in 3 ways:

1) we are mutating less state - just incrementing a single integer,
   rather than lots of rows.
2) the same solution will work exactly the same way for audit logs.
3) it means that automatically comments only become visible to the other
   side when a request is RETURNED or re-SUBMITTED. This means an
   output-checker can add a public comments for the author, but the
   author won't see them until the output-checker is read. And like
   wise, output-checkers won't see the users responses until the user
   re-SUBMITTED the request.

It essentially separates out the PUBLIC/PRIVATE aspect of comments
entirely from the current visibility, and makes the current visibility
based on the request status.

This has some downsides - we now need to know request status to make
decisions, so the responsibility of all that moves to the request
object, which made some templating constructs more awkwar.  But I think
its worth the trade off.