Prevent extraneous alerts stacking up
To reproduce the original issue:

 * As an output checker, approve/reject all files in a release request
   so that the confirmation message "All files reviewed. Your review can
   now be completed." is displayed.

 * In the left-hand tree view, click back and forward between two files
   which updates the page contents without reloading the page.

 * Do something to reload the page (pressing F5 will do).

 * Note the stack of confirmation messages, whose count corresponds to
   the number of times you clicked in the left-hand tree view.

To my mind it's a slightly unusual use of the messages framework to set
messages repeatedly on GET, based on the current state of the system.
Normally, messages are set in response the changes of state, usually via
a POST. But in any case, preventing this messages being set when the
page is requested via HTMX solves the immediate issue.

Closes #464