Record the completed reviewers at the end of a turn
We only want to count file votes from completed reviews when we
determine the overall decision status of a file. However, when we
complete a turn and pass the request back to the author, we need
to show the author what the decisions were at the end of the
previous turn, so we need to know who completed reviews in the
turn that's just ending.