Refactor factories to use BLL to approve/reject files.
Using the DAL allowed us to bypass various checks and restrictions.
Using the BLL instead requires us to make sure that our factories
properly respect these conditions. This is a good thing, but is somewhat
complex.

We can not approve/reject files until the state is SUBMITTED, so we need
to delay that action until then. This meant that the callback pattern of
factories.request_file does not fit, so I added a TestRequestFile class,
and request_file now creates and instance of that. The
create_request_at_status function adds and votes on files using that, at
the right time in the process. The factories API doesn't change, but
the implementation is now quite different.

This required and number of changes exposed a number of test state setup
changes, mostly around making sure we have a fresh instance of the
request, which the BLL requires but the DAL didn't.

One test got particulary complicated due to this change, as setting up
a WITHDRAWN but APPROVED file using the BLL requires use to walk through
more state set up.