Htmx the file votes
After a file vote, we need to replace the 3-way voting buttons to show
the new status. This is the straight out htmx swap. Then we also need to
update the decision and the reviewer's vote to show the right status and
colour - these are done as htmx oob swaps.