Also update the action required message and request buttons after votes
After a changed vote, we not only need to update the vote and its
tree display, but we may also need to update the request buttons
that are enabled/disabled, and the action-required alert. E.g.
if we're in fully-reviewed status, with all files approved, the
"release files" button will be enabled. If a  user changes their
vote to request changes, we need to disable the release files
button and change the alert so it doesn't say that releasing
files is an option.