Fix visible logic when phase is COMPLETE
Previously, our test case didn't get the request state over the line to
where the phase was COMPLETE, so we missed the fact that final comments
might not be shown.

So I changed the big test to reject instead of return, which moves us to
COMPLETE phase, and then added the logic handling correctly to
filter_visible_items.