Fix bug in comment descriptions and class
Previously, whilst in INDEPENDANT phase, all comments would be shown
with the independent description, for researchers or output checkers

Now, that logic only applies to comments from the current round, which
are the only ones subject to blinding.

Additionally, I noticed that blinded comment text didn't make it clear
whether it was a public or private comment, so it now builds on that
information.

Added an explicit test to cover this logic.

Fixes #512