Redirect the coderepo root to project.yaml
The root of a code repo doesn't contain any displayable information,
so it was showing a blank page, which was somewhat confusing to land
on from the source code link in a workspace/request file.

As every linked coderepo should have a project.yaml, and that's the
most likely place a user is going to want to start from when
browsing the source code, we redirect the root to the project.yaml.

There's a bit of fiddliness around the htmx swaps; once in the
code tree, if you click on the root again, we need to make htmx
do a client-side redirect otherwise we don't get the project.yaml
correctly highlighted in the tree as the selected path.

An alternative to this would have been to change the links from the
request/workspace file to go to the project.yaml directly, but then
we'd still get the blank root page when browsing the tree.