Pass the return_url GET param when selecting the coderepo root
When we click on the coderepo root from the tree, the request is
sent via htmx. The view will redirect, but it will have lost the
return_url parameter from the original request, so it loses the
helpful "Back" link in the repo view. So we need to pass it with
the URL for the repo root link in the tree.