Change source code links to project.yaml
The root of a code repo doesn't contain any displayable information,
so it was showing a blank page, which was somewhat confusing to land
on from the source code link in a workspace/request file.

As every linked coderepo should have a project.yaml, and that's the
most likely place a user is going to want to start from when
browsing the source code, we link directly to the project.yaml from
the workspace/request files, instead of to the root.