Prevent resetting a file vote after a review has been submitted
We allow users to vote (approve/request changes) on files, and to
reset those votes. However, after they have submitted their review,
we consider that all files have been voted on, and we use that
to determine the overall decision status of a file. When two
reviewers have submitted reviews, they can discuss and consolidate,
so we need them to be able to change their votes. However, we don't
want them to unset their votes entirely, otherwise the request could
be returned to the author with incomplete decisions on files.

We therefore prevent users from resetting votes after they have
submitted a review in a turn.