Refactor the comment visiblity pills
The language was too long to fit inside the pill.

Given these comment visibilites are only shown to output-checkers, we
just used the "Private", or "Public", and an optional "Blinded" one.

I tried to add tooltips for these with the text description, but they
don't current work properly inside pill components. Thomas is looking at
tooltips, so we can fix that then.  I've left the machinery for
displaying them intact, but have commented out in the template.

I'll add a ticket to renable the tooltips once they work inside a pill.