Refactor the voting urls to simplify template logic
Previously, we were setting the approve/reject/reset urls to either
the relevant URL or None, and using combinations of those to
determine in the template whether the buttons should be shown and
whether they should be disabled or not. This simplifies the logic in
the template and moves it into the view. In the template, we can now
just look up whether the voting buttons should be shown at all, and
their enabled/disabled state.