Move BusinessLayerLogic exceptions to a separate module
We may want to expose these on the object later, but for now
move them out to a) reduce the size of the BusinessLogicLayer class
and b) make some of these exceptions reusable elsewhere. E.g.
User raises an ActionDenied exception which we catch and turn
into a RequestPermissionDenied exception. With exceptions in their
own module, we can just raise the RequestPermissionDenied directly.
This will help with the new separated-out permissions module, where
permission checks can just raise the relevant exception.