Refactor permissions.py into polcies and permissions
We want to extract to a common location all of our checks on
whether a user can do a thing. Some of these checks are
permissions (e.g. a user never has permission to review a
request if they are not an output checker). However, some are
dependent on the state of the request or file that the user is
trying to act on. These change, and so aren't solely user-
dependent permissions. E.g. a user may have permission to
review a request, but can't actually review at this stage because
the request is in RETURNED status, and can only be worked on
by the author.

This splits permissions and policies, where policies are checks
on whether a permitted user can peform an action on a request/file
at this particular time/status.