Refactor request group template context
The request view is mostly construction of a very large template
context, and is quite large and unwieldy.

This commit factors out the construction of the group related template
context into a presenter helper function, and un-flattens the context
dict, putting it all behind a single `"group"` top level key, and
updates the group template to use the new values.