Add view context modal to contents view
This will pop up a diaglog containing the C3 context for the current
files group. I made it a modal so the user doesn't have to navigate away
from the current file.

I added an "inline" in group template to customise the rendering for
this use case a bit. If we're displaying the grup in line:
 a) we don't show the activity
 b) we don't show create comment form.
 c) we add a button to close the modal

We may want to re-visit this later - there are plans to rearrange how we
show activity, and maybe support writing comments from a file page.