Only show request actions on request overview page
This moves the various request action buttons out of the header and onto
the top-level request overview page. The area in the header previously
occupied by the buttons is replaced by a link to the request overview.

This results in some more hoping between pages during the end-to-end
tests, but nothing too complicated.

Closes #472