Replace `pip-compile` command with `compile-reqs`
This follows on from #617 by adding a single `just` command which does
the right thing with respect to compiling requirements files.

There's now no need for a separate `pip-compile` command: it's easy
enough to run specific `pip-compile` commands by hand if you really need
to, and the new command echos what it's doing for transparency.

The command name itself deliberately doesn't include `pip`: its use is
an implementation detail and it's possible we'll swap it for `uv`
further down the line.