Remove ability to customise the id of a request
This had been added in the days before the job-server API had been
changed to accept the ID airlock sent it, and would create its own id.
It's now redundant, and is a bit of functionality that is only in the
private _create_release_request, so we can now remove that in the next
commit.