Remove bll._create_release_request
Originally created to help with test setup, we have now improved the
test factories and other things so that we can use the
`bll.get_or_create_current_request` method. This has the advantage of
removing an alternate path to creating requests, and forcing all request
creations to have valid permissions.

This required fixing a lot of tests' setup data.  For the common case,
create_user will default to having permissions on a workspace called
"workspace", which is the default in many tests. Next commit will
improve this API further.