Separate factories.create_user use cases
Previously, create_user could take a list flat list of workspace names,
or an explicit dict of the full workspace metadata.

This was awkward, and caused some confusion in test setup. So, this
change splits apart those use cases.

The common case, the flat list of workspace names, is still
create_user(). However, that now users a alternate
create_user_with_dict() function under the hood, which is also exposed
to users, and the handful of tests that need to use an explicit dict are
updated to use the new function.