Screenshot the login form in functional tests
Add a helper function for screenshotting an element with some
padding, and screenshot the login form during a test.

My plan is to generate screenshots during the functional tests
to use for the docs. This should keep any screenshots in the docs
up to date with UI changes and styling. It might require some future
changes to the test setup to make the screenshots have a bit more
realistic data in them.