Add How to edit file on request page
Consolidates the update and withdraw file pages into one
"edit file" how-to, which also includes how to change a file
type or group. These two are identical, but are in their own
sections in anticipation of moving a file between groups
becoming a thing that we allow users to do in future.