Move screenshotting functional tests to their own tests
Screenshots were scattered around the functional tests, which made
it difficult to identify where each screenshot used in the docs
came from. They are now all in dedicated tests, which avoid
unnecessary clicking around and asserting content as much as possible.

This does mean that the screenshot tests duplicate quite a lot of the
e2e tests. However, we don't necessarily want to run them by default,
because many of them will change on every run (the request ID will
change, and timestamps of files and activity logs also).

There are some hacky things I experimented with to fix the request
ID and the timestamps, but I think it's probably better that we only
run these tests to update screenshots when we know we want to do so.
The test can run during CI to ensure that there are no actual errors
in it that need to be fixed up.

Not fixing the timestamps has the additional benefit that the
timestamps can serve as a record of when the screenshots were last
updated.