Add scheduled CI task to validate local job-server setup
This required updating the setup scripts to skip the need to have
`SOCIAL_AUTH_*` properly configured.

The local job-server still runs fine w/o it, but it wouldn't be able to
actually log users in from github.

There's a CI jobs that is meant to notify #team-rap slack channel when
this happens, but for reasons I am not sure of, it does seem to work.
Seems like fixing it will mean addressing fact that our slack bot token
setup is undocumented and unclear.