Move visibility logic into its own module
This carves off a small chunk of the BLL logic into its own module for
handling the complexities around vote/comment/audit visibility.

Also moves the main integration test for this logic into its own test
file. This is not exactly a unit test - due the nature of the complex
state manipulation, its more of an integration tests, so its been moved
into that test directory. It is our longest and most complex test, with
a whole support class to aid, so breaking it out makes a certain amount
of sense.