Move AirlockContainer into file_browser_api.py
AirlockContainer defines and abstraction that the file browser API can
use to render and navigate through a tree like structure. The fact that
AirlockContainer was defined in business_logic.py is due to some awkward
import cycles. Those issues have all been fixed, so AirlockContainer can
now be defined in the same module its mainly used.  The Workspace and
ReleaseRequest classes still need to implement it, and mypy will
complain if they don't. But that is all now a concern of the
file_browser_api module, and business_logic does need to care about it
anymore.