Omit docs screenshot tests from coverage
We only run the tests that generate the docs screenshots when we
need them (and not by default with just test-all). Omit them from
coverage so we don't get spurious coverage failures locally
because of the skipped tests. There's nothing critical in these
tests (or the helper functions that only they use) that we need
to ensure coverage for.