Move check for output files to check_can_submit_review policy
Move the check for output files to a policy, instead of just hiding
the submit review button in the UI. This moves the permission
check earlier in the process, which means that a test request created
in PARTIALLY_REVIEWED or REVIEWED status now needs to have at least
one output file. It also means that we don't need to check for
output files when setting the request status, as it will have already
been checked for submitting a review, which is required for changing
status.