Hide help text in the group modal
The filegroup form is shown readonly from the Context button on a
file view. Hide the context/controls instructions here, since no-one
can modify them here.

We keep the instructions on the group page, even if the form is readonly
(either the user is a reviewer, or the request isn't in an
editable-by-author
status, as they'll still be a useful reminder.