Add a renderer for log files that can render ansi colours
Convert the text of the log file to HTML using ansi2html, which
converts ANSI colour codes to css classes so we get the colour
formatting from the original log.

It produces a full HTML file, which isn't what we want to display.
We just extract the <pre></pre> tag which contains the log content,
and the <style> which contains the relevant inline css.