Fix bug in CSV filtering
Previously, our datatable <table> started out with a "hidden" class,
which was removed when the datatable was initialised. This prevented the
appearance of unformatted tables before the nicely formatted datatable
was ready.

However, when a filter matches nothing and is then reset, simple
datatables refreshes the original table element, which was reinstating
its hidden attribute, so we ended up with a blank page.

We now wrap the table in another containing div, and apply the hidden
class to that instead. It has the same effect in hiding the table until
it's initialised, but isn't affected by the refresh after a removed
unmatching filter.