Handle django control flow exceptions better in tracing.
By default, if an exception is raised during an operation being measured
by an otel span, it will mark the span as being in an error state, as
well as record the exception details.

However, django uses exceptions for non-error control flows, such as
raising Http404, or PermissionDenied. This are not error, but due the
above defaults, they show up in our telemetry as errors, and we are
alerted to 404s and 403s, which we do not what.

Given that the `instrument` decorator's primary use case is decorating
django view methods, this changes the default behaviour to set
set_status_on_exception=False.  Actual error exceptions will continue
to bubble up, and the root span will have its status set to error, but
the child spans generated by `instrument` will not.