Ensure that tests don't try to call real endpoints
In setting.py, AIRLOCK_API_ENDPOINT defaults to the real job-server
endpoint, so if it isn't set in a .env file when the tests run, and
a test isn't mocking the api calls properly, it could attempt to call
the real endpoints, which isn't ideal.

This just mocks the setting for all tests. It does it in the same
fixture as the existing autouse fixture that overrides file storage
otherwise the two conflict.