Only release file and audit if it hasn't already been uploaded
"Release file" indicates only an intention by the user to release
a file. If an upload failed and a user (the same or different) is
attempting to re-try the release, we only want to update the release
metadata and audit event for the files that haven't already been
uploaded, and which we're just re-setting the upload
attempts so we can attempt to upload again.