Remove last vestiges of naked "project" field in user API
AFAICS, we do not use this in production code at all, but there were
a few places where it was used in tests, which this cleans up.

The default create_user test factory does not set it, just some hand
created tests.

The motivation is to clean up the API, as there is a note in job-server
side that we need it for b/w compat, which I think is no longer true.

https://github.com/opensafely-core/job-server/blob/main/jobserver/api/releases.py#L588