Don't set the request status to released in bll.release_files
When a user releases files, the status is set to approved, and the
release is created on job-server. The release isn't actually done
until all the files are uploaded, so we remove the status change
when the release is started. The task runner will set status to
released when all files are uploaded.