Move uploaded_file_count to an included html snippet
We're going to want to replace this snippet with htmx so we can
update the file upload progress.

Also adds upload_in_progess as a variable in the request view
context as we're going to need this more than once and it'll
avoid repeatedly calling the method which hits the db.