Poll for changes to the uploaded files count
If a request has uploads in progress, we poll every second for updates
and replace the uploaded_files_count element.

If on an update we find that the request is no longer actively
uploading,
we reload the page to update the various other elements that might need
to change, and we stop polling.