Separte last_fresh from auth api
Currently airlock/login_api.py handles the job server auth api data, as
well as adding prod/dev specific last_refresh, merging both into
a single dict object that is the User.

For the incoming db User object, I want to more formally separate what
is API data, and what is local airlock db state (e.g.
last_refresh).

So, this change simplifies the login_api to just return the API data,
and makes setting last_refresh timestamp the job of the middleware that
checks it.

This did mean that dev users with be re-authenticated, so we support
that in get_user_authz by always successfully re-authenticating the
currently logged in user.