Update the definition of can_be_released()
Now that we will retry uploads until they succeed, there's no need
to allow users to release when the request is in approved status.
Approved can now properly be a system-owned status (as it was before).

This also allows us to simplify the logic in the request view for
when the release-files button is shown.