Update the definition of upload_in_progress()
Now that we're retrying all uploads forever, a request is always in
upload-in-progress status when it's approved, even if all of its
uploads are actually done. The file uploader isn't done until it's
also changed the request status to RELEASED.

We could just check request.status == APPROVED, but this is more
descriptive in terms of what the uploader's doing.