Ignore empty groups with missing context and controls
We prevent users from submitting requests with groups that are
missing context and controls. However, if some of those groups are
empty (i.e. have no output or supporting files), we shouldn't care
if they're missing context/controls. Note that the group still exists
and should be displayed, even if it has no files. It only gets to
this state if all files have been withdrawn, and it would be
reasonable for a user to still want to see the group details (e.g.
activity log, previously entered context/controls).