Add a user migration command
This should populate the db with the details of users who have done
anything in the system. This should ensure their data is present in
future, e.g. fullname, even if they never log in again.

I have not tested it, as it needs real data/API_TOKEN and is going to be
short lived.