Replace wait_for_timeout with wait for htmx classes
Playwright docs warn that wait_for_timeout is flaky
https://playwright.dev/python/docs/api/class-frame#frame-wait-for-timeout

Replace it with a wait for the htmx class that's added and then
removed during an htmx operation.
https://htmx.org/docs/#request-operations

Also wait for htmx to be done in the find_and_click function. Not all
of these clicks will involve htmx, but lots of them do. Note that we
could wait for htmx-added counts of 1 and then 0, which would verify
that an htmx swap happened - but it doesn't always pick up the
1 count, even when it's definitely an htmx operation, so I think
sometimes it's too fast for playwright; in any case, this means that we
can add it in the find_and_click method without being too bothered
about whether it's htmx or not, we just make sure there's no
htmx-swapping still going on.