Add temporary live backfill of user data
We are not currently using the User db object, but this change
pre-emptively populates the table.

We will run the populate_users command manually, but having this here
will ensure nothing breaks if new users happen to log on in between
running it and deploying the version tht uses it.