Update find_and_click helper
Previously this helper called expect(locator).to_be_visible() before
click. The intention was to make it fail earlier if it wasn't visible,
because we set a longer-than-default timeout (15s) and click() waits
not only on visibility but also for the element to be enabled, stable
and ready to receive events. This was useful when we were first
building the functional tests, but it shouldn't be needed now, and
just in case it's contributing to the flakiness, let's just call
click() directly and let it do all the waiting it wants to.