Add tooltips to the overall decision display
Add an explanation of what the overal decision display means. I
haven't made this take account of the user's current vote, because
it would be quite a lot of additional logic; I think this should
be enough to explain why e.g. you're an output checker who's
clicked Approve and is wondering why the overall decision is still
Incomplete. It also means the tooltips should still make sense
whether you're looking at the request as an output checker or an
author.