Don't reload the page after changing status to released
The htmx polling is still going on at this point, so there's
no need to reload the page, and possibly doing so is causing
the occasional conflict (a "maybe frame was detached?" error
on the page.reload() call). We just check that the header
status has changed to RELEASED before taking the screenshot.