Switch from username string to User object in BLL/DAL
Previously, whenever we needed to track a user, e.g. in a request,
comment, vote, audit or other object, we used the username string. This
was in part that we only had a User instance for logged in user, not
other users we need to record or display.

Now, we have a formal User object that is persisted, this change
switches to using that everywhere in the BLL/DAL API. So,
`ReleaseRequest.author` is now type User rather than string, etc.

This means that wherever we interact with User, we have the full
information available, in particular their fullname. This has made the
API more consistent, and allows us to more easily extend User properties
in future.

The local_db DAL implementation still just persists the user id, at it
is not responsible for the User metadata data. The BLL is reponsible for
looking up the User object using this user id as part of
to_dict/from_dict interface between the two.