Fix workspace name/repo name mixup
This resulted in the code browser generating the wrong URLs in its file
menu which meant that the initial load of the `project.yaml` file
worked, but clicking any of the other filenames would silently fail (the
HTMX call would 403 but no error would be displayed to the user).

In cases where the repo name happened to match the workspace name things
would work fine, which is why we didn't spot this locally.

Closes #845