Update telemetry with new user details
This fixes a missed check in the services instrument to be
user.is_authenticated, rather then user existing, as with
django.contrib.auth, the user always exists.

Also, track the user_id and username separately, to match the code, and
to be ready for when user_id changes.

Additionally, rather than have an emtpy string when there is no user,
use the string "anonymous" as it is unambiguous when veiwing the
telemetry, and matches the way django.auth.contrib works