Require public comment for requested changes in order to return
If a request is being returned with changes requested, there must
be a public (i.e. visible to the requester) comment on every file
group with changes-requested file, so the requester has information
about what changed are requested.

Note that we require comments but still allow the request to be
returned if the decision is conflicted; at some point we might want
to require consensus of reviews before returning, but that's for another
time.

Applies to fully reviewed requests only; requests can be returned at
any point. If it's being returned before independent review is complete,
it's likely to be because the requester made a mistake and asked for it
to be reviewed so they could update and resubmit. In this case, we
don't require the reviewer to add public comments to all filegroups
with changes requested.