Add url and view for requesting changes for a filegroup
This view just loops through all the output files in a filegroup
and votes to request changes on any file that hasn't already been
voted on.  If the file is approved, it's left as approved. If it's
already had changes requested, we don't vote again.

There's a very tiny chance that a request could changes status
during this process, so some files are changed and some fail to
change. We keep track of that and report it back to the user
rather than raising the exceptions immediately.