Redirect to group after deleting all files from a directory
Withdrawing files from a pending request deletes the files. If we
withdraw all files from a directory (using the multiselect), we
can't redirect back to the directory, because it's no longer a
valid url, so we redirect to the group instead. This is consistent
with the behaviour when withdrawing a single file from a pending
request.