Use the file uploader in screenshot tests
This point in the test has seen the occasional "db locked" error,
which could possibly be due to the htmx polling interfering with the
manual bll calls. Calling the run_file_uploader management command
makes for simpler code in the test and more closely mimics the
way the uploads work.