Add some extra expects around the iframe locator in functional tests
We're still getting some occasional failures on the iframe content
in the functional tests. This just adds an additional expect() around
the locator itself to try and figure out whether it's the locator
itself that hasn't been found (i.e. the iframe container in the
contents) or the loaded file content.