Cache templates based on their content
We don't want to cache templates with their mtime, because this
time will update when we redeploy in production, even if the
template content hasn't changed. Django caches templates by default,
so retrieving the template source content using the template loader
is cheap.

We then don't need to check the last-modified time in serve_file(),
because the renderer etag incorporates the file's last modified
time as well as the template content. If one of these has changes, the
etag will change, and we want to reload, even if the last modified
time (based on the file only) hasn't changed.