Reload the template if it has been modified
Django by default caches templates, even in development.
The RendererTemplate classes are defined on the Renderer classes, so
the templates themselves are never reloaded even if a template
file changes. Django's autoreload watches for template changes and
clears the template cache, but in order to pick up those changes
we need to ensure the RendererTemplate is reloaded.

Note that the template cache is in-memory, so starting and stopping the
development server would clear it. However, if you've changed a template
and reloaded the page in the browser, the eTag and LastModified headers
will be updated, so even when you start/stop the server, the changes
don't get picked up.