Use the file uploader to upload files during functional tests
This point in the test has seen the occasional "db locked" error, which could possibly be due to the htmx polling interfering with the manual bll calls. Calling the run_file_uploader management command makes for simpler code in the test and more closely mimics the way the uploads work.

(A similar update - in https://github.com/opensafely-core/airlock/pull/900 - was done
for the screenshot functional tests; which also encountered the "db locked" error, somewhat more frequently, and haven't (so far) since the change was made)