Make state diagram generation a django manage command
In django 5.2 there is some extra output for the shell command
that prints information about automatic imports
https://github.com/django/django/blob/56e23b2319cc29e6f8518f8f21f95a530dddb930/django/core/management/commands/shell.py#L198

This means if we just send the output of the statemachine script to
a file, we get that output in the file too.  This turns it into a
standard manage command that writes to the file instead.