Use a filesystem db in tests
Django by default uses an in-memory database during tests. We can't
set WAL mode on an in-memory database, so the functional tests that
have htmx polling and also run the file uploader occasionally hit a
"database table is locked" error when the polling tries to read at
the same time as the file uploader writes. We can use a filesystem
db instead by setting a test db name in the DATABASES settings.