Add basic test framework for running gunicorn.
To test guniocrn, we need to be able to run it in the background, and
send requests to it.  This has subtley complexity, I have pulled in
a solution pattern I have used previously, providing a context manager
that runs gunicorn in subprocess over a unix socket, and DTRT if it
fails to come up.

We use a unix socket because it is isolated to the specific test.
Running gunicorn on a TCP port avoids the need to coordinate TCP ports.
This does mean bringing in httpx so we can make requests over a unix
socket.